Modified to use new wrapper code.
git-svn-id: https://192.168.3.139:8443/svn/Medical/trunk@965 956a8c8d-885e-cb46-a0b9-a2f355b4c364

+++ Common/Controller/ImageRenderer.cs
@@ -220,7 +220,7 @@ namespace Medical
             BitmapData bmpData = bitmap.LockBits(new Rectangle(new Point(), bitmap.Size), ImageLockMode.WriteOnly, bitmap.PixelFormat);
             unsafe
             {
-                PixelBox pixelBox = new PixelBox(0, 0, (uint)bmpData.Width, (uint)bmpData.Height, format, bmpData.Scan0.ToPointer());
+                PixelBox pixelBox = new PixelBox(0, 0, bmpData.Width, bmpData.Height, format, bmpData.Scan0.ToPointer());
                 renderTexture.copyContentsToMemory(pixelBox, RenderTarget.FrameBuffer.FB_AUTO);
             }
             bitmap.UnlockBits(bmpData);
@@ -322,7 +322,7 @@ namespace Medical
                         BitmapData bmpData = pieceBitmap.LockBits(new Rectangle(new Point(), pieceBitmap.Size), ImageLockMode.WriteOnly, pieceBitmap.PixelFormat);
                         unsafe
                         {
-                            PixelBox pixelBox = new PixelBox(0, 0, (uint)bmpData.Width, (uint)bmpData.Height, format, bmpData.Scan0.ToPointer());
+                            PixelBox pixelBox = new PixelBox(0, 0, bmpData.Width, bmpData.Height, format, bmpData.Scan0.ToPointer());
                             renderTexture.copyContentsToMemory(pixelBox, RenderTarget.FrameBuffer.FB_AUTO);
                         }
                         pieceBitmap.UnlockBits(bmpData);
